+++ docs/pages/_app.js
@@ -1,6 +1,13 @@
+import { Analytics } from '@vercel/analytics/react';
+
 import 'nextra-theme-docs/style.css';
 import '../styles/global.scss';
 
 export default function Nextra({ Component, pageProps }) {
-  return <Component {...pageProps} />
+  return (
+    <>
+      <Component {...pageProps} />
+      <Analytics />
+    </>
+  );
 }